v11

+++ accounts/views.py
@@ -3,7 +3,7 @@ from django.contrib.auth import login, authenticate, logout
 from django.contrib import messages
 from django.shortcuts import render, redirect
 from .forms import SignUpForm
-from django.http import HttpRequest, HttpResponse
+from django.http import HttpRequest, HttpResponse, HttpResponseNotFound
 from .backend import process_ppt
 from django.conf import settings
 import os
@@ -19,7 +19,9 @@ from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 from django.utils.encoding import force_bytes
 from . tokens import generate_token
 from pymongo import MongoClient
+import mimetypes
 from django.contrib.auth.hashers import make_password
+
 
 client = pymongo.MongoClient(settings.MONGODB_URI)
 db = client[settings.MONGODB_NAME]
@@ -27,32 +29,50 @@ collection = db['enhanced_ppt']
 
 def upload_ppt(request):
     if request.method == 'POST':
-        ppt_file = request.FILES.get('ppt_file')
-        if ppt_file:
-            try:
-                # Save the uploaded file to a temporary location
-                with open('temp.pptx', 'wb') as f:
-                    for chunk in ppt_file.chunks():
-                        f.write(chunk)
-                
-                # Process the uploaded PowerPoint file
-                enhanced_ppt_path = process_ppt('temp.pptx')
-                
-                # Save the enhanced PowerPoint presentation to MongoDB
-                with open(enhanced_ppt_path, 'rb') as f:
-                    enhanced_ppt_data = f.read()
-                    ppt_doc = {'name': 'enhanced_presentation.pptx', 'data': enhanced_ppt_data}
-                    collection.insert_one(ppt_doc)
+        # Retrieve the username from the session
+        username = request.session.get('username')
 
-                # Delete the temporary file
-                os.remove('temp.pptx')
-                
-                ppt_id_str = str(ppt_doc['_id'])  # Convert ObjectId to string
+        # Check if the username is present in the session
+        if username:
+            ppt_file = request.FILES.get('ppt_file')
+            
+            if ppt_file:
+                try:
+                    # Save the uploaded file to a temporary location
+                    with open('temp.pptx', 'wb') as f:
+                        for chunk in ppt_file.chunks():
+                            f.write(chunk)
+                    
+                    # Process the uploaded PowerPoint file
+                    enhanced_ppt_path = process_ppt('temp.pptx', username)
 
-                return redirect('preview_ppt', ppt_id=ppt_id_str)
-            except Exception as e:
-                return HttpResponse(f"An error occurred: {str(e)}", status=500)
+                    # Delete the temporary file
+                    os.remove('temp.pptx')
+                    
+                    
+                except Exception as e:
+                    return HttpResponse(f"An error occurred: {str(e)}", status=500)
+        else:
+            # If username is not present in the session, redirect to the login page
+            return redirect('login')
     return render(request, 'page.html')
+
+def view_past_ppts(request):
+        username = request.session.get('username')
+
+        # Check if the username is present in the session
+        if username:
+            # Retrieve all PowerPoint presentations where the user attribute matches the username
+            past_ppts = collection.find({'user': username})
+            modified_ppts = []
+            for ppt in past_ppts:
+                ppt['str_id'] = str(ppt['_id'])
+                modified_ppts.append(ppt)
+
+            return render(request, 'past.html', {'past_ppts': modified_ppts})
+
+        else:
+            return redirect('login')
 
 def entry(request: HttpRequest):
     return render(request, 'entry.html')
@@ -60,6 +80,7 @@ def entry(request: HttpRequest):
 client = MongoClient(settings.MONGODB_URI)
 db = client[settings.MONGODB_NAME]
 user_collection = db['users']
+collection = db['enhanced_ppt']
 
 def signup(request):
     if request.method == 'POST':
@@ -104,15 +125,32 @@ def login(request):
         # Check if user exists in MongoDB
         user_data = user_collection.find_one({'username': username, 'password': password})
         if user_data:
-            # # Authenticate user in Django
-            # user = authenticate(request, username=username, password=password)
-            # if user is not None:
-                # login(request, user)
+                request.session['username'] = username
                 messages.success(request, 'You are now logged in.')
-                print("hello")
-                return redirect('upload_ppt')  # Replace 'home' with your desired redirect URL after login
+                return redirect('entry')  
 
         messages.error(request, 'Invalid username or password.')
         return redirect('login')
 
     return render(request, 'login.html')
+
+
+def download_presentation(request, presentation_id):
+    presentation = collection.find_one({'_id': ObjectId(presentation_id)})
+
+    if presentation:
+        # Retrieve presentation data and metadata
+        presentation_data = presentation['data']
+        filename = presentation['name']
+
+        # Determine the MIME type of the presentation
+        content_type, _ = mimetypes.guess_type(filename)
+        if not content_type:
+            content_type = 'application/octet-stream'
+
+        # Set response headers for file download
+        response = HttpResponse(presentation_data, content_type=content_type)
+        response['Content-Disposition'] = f'attachment; filename="{filename}"'
+        return response
+    else:
+        return HttpResponseNotFound('Presentation not found')
